refactor: 💡 add type module and ts
