+++ package.json
@@ -4,6 +4,7 @@
 	"author": "s-hirano-ist",
 	"description": "Private pages and admin tools for s-hirano-ist.",
 	"packageManager": "pnpm@9.15.4",
+	"type": "module",
 	"scripts": {
 		"dev": "next dev",
 		"build": "next build",
