+++ sentry.client.config.ts
@@ -2,7 +2,7 @@
 // The config you add here will be used whenever a users loads a page in their browser.
 // https://docs.sentry.io/platforms/javascript/guides/nextjs/
 
-import { env } from "@/env.mjs";
+import { env } from "@/env";
 import {
 	captureConsoleIntegration,
 	init,
