+++ sentry.server.config.ts
@@ -2,7 +2,7 @@
 // The config you add here will be used whenever the server handles a request.
 // https://docs.sentry.io/platforms/javascript/guides/nextjs/
 
-import { env } from "@/env.mjs";
+import { env } from "@/env";
 import * as Sentry from "@sentry/nextjs";
 
 Sentry.init({
