+++ src/features/auth/utils/auth.config.ts
@@ -1,4 +1,4 @@
-import { env } from "@/env.mjs";
+import { env } from "@/env";
 import type { NextAuthConfig } from "next-auth";
 import Auth0 from "next-auth/providers/auth0";
