+++ src/features/image/actions/generate-url.test.ts
@@ -1,5 +1,5 @@
 import { ERROR_MESSAGES, SUCCESS_MESSAGES } from "@/constants";
-import { env } from "@/env.mjs";
+import { env } from "@/env";
 import { auth } from "@/features/auth/utils/auth";
 import { minioClient } from "@/minio";
 import { Session } from "next-auth";
