+++ src/prisma.ts
@@ -1,6 +1,6 @@
 // In case of error of prisma on development environment
 // https://www.prisma.io/docs/orm/more/help-and-troubleshooting/help-articles/nextjs-prisma-client-dev-practices
-import { env } from "@/env.mjs";
+import { env } from "@/env";
 import { PrismaClient } from "@prisma/client";
 // import { PrismaClient } from "@prisma/client/edge"; // FIXME: for edge
 import { withAccelerate } from "@prisma/extension-accelerate";
