+++ tsconfig.json
@@ -22,7 +22,7 @@
 		"**/*.ts",
 		"**/*.tsx",
 		".next/types/**/*.ts",
-		"vitest.config.mts"
+		"vitest.config.ts"
 	],
 	"exclude": ["node_modules", "script/**/*", "README.md"]
 }
