+++ vitest.workspace.ts
@@ -4,7 +4,7 @@ import { defineWorkspace } from "vitest/config";
 
 // More info at: https://storybook.js.org/docs/writing-tests/vitest-plugin
 export default defineWorkspace([
-	"vitest.config.mts",
+	"vitest.config.ts",
 	{
 		plugins: [
 			// See options at: https://storybook.js.org/docs/writing-tests/vitest-plugin#storybooktest
